Add separate environment variables for C and C++ compilers
Before this commit, we used LIS_CC to specify the C compiler, and it
was used for both compiling C and C++ source code.

This commit adds LDT_CPPC, LIS_CPPC, and LVT_CPPC environment variables
for specifying the C++ compiler.

The LDT_CC, LIS_CC, and LVT_CC environment variables are still used
to specify the C compiler.